pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-api-python-client>=2.164.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
